requirements.txt
langgraph>=0.2.0
langchain>=0.3.0
langchain-anthropic>=0.3.0
python-dotenv>=1.0.0
langchain-groq>=0.1.0
groq>=0.9.0
anthropic>=0.40.0
pandas>=2.0.0
numpy>=1.26.0
pydantic>=2.0.0
docker>=7.0.0
tqdm>=4.66.0
evalplus>=0.3.0
bandit>=1.7.0
coverage>=7.0.0
openai>=1.0.0
